Crate graphsync

Modules

libp2p

proto

Structs

ExtensionData
A name/data pair for a GraphSync extension.

GraphSyncMessage
Contains all requests and responses

GraphSyncRequest
Struct which contains all request data from a GraphSyncMessage.

GraphSyncResponse
Struct which contains all response data from a GraphSyncMessage.

NewRequestPayload
The data associated with a new graphsync request.

Enums

Payload
Defines the data associated with each request type.

ResponseStatusCode
Status code returned for a GraphSync Request.

Constants

EXTENSION_DO_NOT_SEND_CIDS
Tells the responding peer not to send certain blocks if they are encountered in a traversal, see https://github.com/ipld/specs/blob/master/block-layer/graphsync/known_extensions.md

EXTENSION_METADATA
Provides response metadata for a GraphSync request, see https://github.com/ipld/specs/blob/master/block-layer/graphsync/known_extensions.md

MAX_BLOCK_SIZE
The maximum combined block size for a single message.

Type Definitions

ExtensionName
Name for a GraphSync extension.

Extensions
Represents the data attached as extensions to the requests.

Priority
Priority for a GraphSync request.

RequestID
Unique identifier for a GraphSync request.